Added Buttons to main page
Buttons are labeled and color coded. I changed the opacity of the Discard and add to hand buttons. The intention is that these buttons are disabled while there is not a card currently being displayed. Once a card is drawn the draw card button will be disabled until the card is added to the hand or discarded.